Revamped create comments reducer
